+++ core/src/main/java/eu/unicore/ucc/actions/job/Allocate.java
@@ -1,23 +1,26 @@
 package eu.unicore.ucc.actions.job;
 
 import java.util.Arrays;
+import java.util.HashMap;
+import java.util.Map;
 
 import org.apache.commons.cli.Option;
 
-import eu.unicore.client.Endpoint;
 import eu.unicore.client.Job;
-import eu.unicore.client.core.AllocationClient;
+import eu.unicore.client.Job.Type;
+import eu.unicore.client.core.JobClient;
+import eu.unicore.client.core.JobClient.Status;
 import eu.unicore.ucc.UCC;
 import eu.unicore.ucc.actions.ActionBase;
 import eu.unicore.ucc.runner.Runner;
 import eu.unicore.ucc.util.UCCBuilder;
 
 /**
- * Runs a single command through UNICORE
+ * Allocates resources on the batch system via UNICORE
  * 
  * @author schuller
  */
-public class Exec extends ActionBase {
+public class Allocate extends ActionBase {
 
 	protected Runner runner;
 
@@ -29,21 +32,10 @@ public class Exec extends ActionBase {
 	protected String siteName=null;
 
 	/**
-	 * request a particular login node
-	 */
-	protected String loginNode=null;
-
-	/**
 	 * whether to actually submit the job - if <code>false</code>, brokering etc will
 	 * be performed but no job will be submitted 
 	 */
 	protected boolean dryRun=false;
-
-	/**
-	 * whether to keep the finished job - if <code>false</code>, 
-	 * the job will be deleted when done
-	 */
-	protected boolean keep=false;
 
 	protected String[] tags;
 
@@ -53,10 +45,10 @@ public class Exec extends ActionBase {
 	protected boolean asynchronous;
 
 	/**
-	 * Allocation job URL
+	 * resources name / value pairs
 	 */
-	protected String allocation=null;
-
+	private final Map<String,String>resourceRequests=new HashMap<>();
+	
 	@Override
 	protected void createOptions() {
 		super.createOptions();
@@ -66,20 +58,6 @@ public class Exec extends ActionBase {
 				.argName("Site")
 				.hasArg()
 				.required(false)
-				.build());
-		getOptions().addOption(Option.builder(OPT_LOGIN_NODE)
-				.longOpt(OPT_LOGIN_NODE_LONG)
-				.desc("Login node to use")
-				.argName("LoginNode")
-				.hasArg()
-				.required(false)
-				.build());
-		getOptions().addOption(Option.builder(OPT_ALLOCATION)
-				.longOpt(OPT_ALLOCATION_LONG)
-				.desc("Allocation URL")
-				.required(false)
-				.argName("Allocation")
-				.hasArg()
 				.build());
 		getOptions().addOption(Option.builder(OPT_BROKER)
 				.longOpt(OPT_BROKER_LONG)
@@ -91,14 +69,9 @@ public class Exec extends ActionBase {
 				.desc("Dry run, do not submit the job")
 				.required(false)
 				.build());
-		getOptions().addOption(Option.builder(OPT_KEEP)
-				.longOpt(OPT_KEEP_LONG)
-				.desc("Don't remove finished job")
-				.required(false)
-				.build());
 		getOptions().addOption(Option.builder(OPT_MODE)
 				.longOpt(OPT_MODE_LONG)
-				.desc("Run asynchronous, don't wait for finish, don't get results.")
+				.desc("Only submit, don't wait for the allocation to start running.")
 				.required(false)
 				.build());
 		getOptions().addOption(Option.builder(OPT_TAGS)
@@ -110,63 +83,67 @@ public class Exec extends ActionBase {
 				.build());
 	}
 
+	protected void readResources(){
+		resourceRequests.clear();
+		int length=getCommandLine().getArgs().length;
+		if(length<1)return;
+		for(int i=1; i<length; i++){
+			String p=getCommandLine().getArgs()[i];
+			String[]split=p.split("=");
+			String key=split[0];
+			String value=split[1];
+			verbose("Have resource request: "+key+"="+value);
+			resourceRequests.put(key, value);
+		}
+		//unit testing use
+		lastParams=resourceRequests;
+	}
 	@Override
 	public String getName() {
-		return "exec";
+		return "allocate";
 	}
 
 	@Override
 	public String getArgumentList(){
-		return "[<commands>]";
+		return "allocate [<options>] [resource1=value1 resource2=value2 ...]";
 	}
 
 	@Override
 	public String getSynopsis(){
-		return "Runs a command through UNICORE. " +
-				"The command will not be run through a remote queue, but on the cluster login node. " +
-				"The command and its arguments are taken from the UCC command line." +
-				"UCC will wait for the job to finish and print standard output and error to the console.";
+		return "Allocates resources through UNICORE. ";
 	}
 
 	@Override
 	public String getDescription(){
-		return "run a command through UNICORE";
+		return "allocate resources through UNICORE";
 	}
 	@Override
 	public String getCommandGroup(){
 		return "Job execution";
 	}
 
-
 	@Override
 	public void process(){
 		super.process();
 		siteName=getCommandLine().getOptionValue(OPT_SITENAME);
-		allocation = getCommandLine().getOptionValue(OPT_ALLOCATION);
-		if(allocation!=null && siteName!=null) {
-			throw new IllegalArgumentException("Cannot have both 'allocation' and 'sitename' arguments.");
-		}
-		loginNode=getCommandLine().getOptionValue(OPT_LOGIN_NODE);
 		dryRun=getBooleanOption(OPT_DRYRUN_LONG, OPT_DRYRUN);
 		verbose("Dry run = "+dryRun);
-		keep=getBooleanOption(OPT_KEEP_LONG, OPT_KEEP);
-		verbose("Delete job when done = "+!keep);
 		asynchronous=getBooleanOption(OPT_MODE_LONG, OPT_MODE);
 		verbose("Asynchronous processing = "+asynchronous);
 		tags = getCommandLine().getOptionValues(OPT_TAGS);
 		if(tags!=null) {
 			verbose("Job tags = " + Arrays.deepToString(tags));
 		}
-		initBuilder(getCommandLine().getArgs());
+		readResources();
+		initBuilder();
 		run();
 	}
 
-	protected void initBuilder(String[] args){
+	protected void initBuilder(){
 		try{
 			builder = new UCCBuilder(registry, configurationProvider);
 			builder.setProperty("Output",output.getAbsolutePath());
 			builder.setProperty("IDLocation",output.getAbsolutePath());
-			builder.setProperty("KeepFinishedJob", String.valueOf(keep));
 			builder.setProperty("DetailedStatusDisplay", "true");
 			if(tags!=null&&tags.length>0) {
 				builder.addTags(tags);
@@ -175,14 +152,9 @@ public class Exec extends ActionBase {
 				builder.setProperty("Site", siteName);
 			}
 			Job job = new Job(builder.getJSON());
-			if(args.length == 1)throw new IllegalArgumentException("Must specify a command");
-			// first arg is command
-			job.executable(args[1]);
-			job.run_on_login_node(loginNode);
-			if(args.length > 1 ){
-				for(int i=2; i<args.length;i++){
-					job.arguments(args[i]);
-				}
+			job.type(Type.ALLOCATE);
+			for(String rName: resourceRequests.keySet()) {
+				job.resources().other(rName, resourceRequests.get(rName));
 			}
 		}catch(Exception e){
 			error("",e);
@@ -192,7 +164,7 @@ public class Exec extends ActionBase {
 
 	protected void run(){
 		runner = new Runner(registry,configurationProvider,builder);
-		runner.setAsyncMode(asynchronous);
+		runner.setAsyncMode(true);
 		runner.setBriefOutfileNames(true);
 		runner.setOutputToConsole(true);
 		runner.setDryRun(dryRun);
@@ -202,23 +174,28 @@ public class Exec extends ActionBase {
 			brokerName = "LOCAL";
 		}
 		runner.setBroker(UCC.getBroker(brokerName));
-		if(allocation!=null) {
-			try {
-				AllocationClient ac = new AllocationClient(new Endpoint(allocation),
-						configurationProvider.getClientConfiguration(allocation),
-						configurationProvider.getRESTAuthN());
-				runner.setSubmissionService(ac);
-			}catch(Exception ex) {
-				throw new RuntimeException(ex);
-			}
-		}
 		try{
 			runner.run();
-		}catch(RuntimeException ex){
+			if(!asynchronous) {
+				// make sure job is "RUNNING"
+				JobClient job = runner.getJob();
+				verbose("Allocation job is "+job.getStatus());
+				do {
+					if(runner.getJob().getStatus().ordinal()>=Status.RUNNING.ordinal()) {
+						break;
+					}
+					Thread.sleep(2000);
+				}
+				while(true);
+				verbose("Allocation job is "+job.getStatus());
+			}
+		}catch(Exception ex){
 			runner.dumpJobLog();
-			error("Failed to execute remote command", ex);
+			error("Failed to allocate resources", ex);
 			endProcessing(ERROR);
 		}
 	}
 
+	// for unit testing
+	public static Map<String,String>lastParams;
 }