+++ distribution/CHANGES.md
@@ -6,9 +6,13 @@ UCC Changelog
 
 **JAVA VERSION NOTE** This release requires Java 11 or later!
 
-Version 10.0.0 (released Jan dd, 2024)
+Version 10.0.0 (released Feb dd, 2024)
 --------------------------------------
- - fix: shell: OAuth authentication fails when refresh token is present, but expired
+ - new feature: "allocate" command for creating an allocation
+ - 'shell': fix: OAuth authentication fails when refresh token is present, but expired
+ - 'list-jobs': improved job list when using the "--long" option
+ - 'issue-token': show subject when inspecting the token
+ - 'exec': new option "--allocation" for running command in an existing allocation
  - update to UNICORE 10.0.0 base libs
 
 Version 9.3.0 (released Sept 15, 2023)